Timer 1 interrupts working in an ideal case.
Timer 1 will be used to turn on the TRIAC. In this example, Timer 2 is
set to interrupt at 120 Hz simulating the number of times the AC line
crosses zero.

Timer 2 disables the TRIAC, and resets and enables Timer 1. When the
Timer 1 interrupt occurs, it switches on the on-board LED and disables
itself.

This example uses a look-up table for determining how long to wait based
on the desired power.