Refactored the code to use the actual module.
Timer 2 has been removed and replaced with an input interrupt from the
TRIAC module. It interrupts on the RISING edge of the Zero Crossing
line. Due to noise on the AC line, there are usually two successive
interrupts here.

The built-in LED has been switched to the input of the TRIAC.